requirements.txt
opencv-python>=4.8.0
numpy>=1.21.0
matplotlib>=3.5.0
scikit-image>=0.19.0
tensorflow>=2.10.0
torch>=1.12.0
torchvision>=0.13.0
Pillow>=9.0.0
PyQt5>=5.15.0
imageio>=2.25.0
scipy>=1.9.0
pandas>=1.5.0
seaborn>=0.11.0
tqdm>=4.60.0
colorama>=0.4.0
requests>=2.31.0
lpips>=0.1.4
streamlit>=1.28.0
streamlit-drawable-canvas>=0.9.0
plotly>=5.15.0
streamlit-option-menu>=0.3.6
